refactor: Separate struct MyLine from MyPoint.swift
